[ADD/REMOVE/IMPROVE] Added Doxygen docs, removed rest-lavalink and micellaneous, and improved code.
Now Coglink finally have a documentation, and files like rest-lavalink and micellaneous were removed, because they were organized in new files like information, track and network. Internal functions were improved and some struct names were changed to be "better".
