+++ include/lavalink-internal.h
@@ -15,7 +15,20 @@ struct io_poller;
 #define __COGLINK_DELETE_REQ 2
 #define __COGLINK_PATCH_REQ 3
 
-int __coglink_performRequest(struct lavaInfo *lavaInfo, int requestType, int additionalDebuggingSuccess, int additionalDebuggingError, char *path, int pathLength, int useV3Path, char *body, long bodySize, struct requestInformation *res, int getResponse, CURL *reUsedCurl);
+struct __coglink_requestConfig {
+  int requestType;
+  _Bool additionalDebuggingSuccess;
+  _Bool additionalDebuggingError;
+  char *path;
+  int pathLength;
+  _Bool useV3Path;
+  char *body;
+  long bodySize;
+  _Bool getResponse;
+  CURL *usedCURL;
+};
+
+int __coglink_performRequest(struct lavaInfo *lavaInfo, struct requestInformation *res, struct __coglink_requestConfig *config);
 
 int __coglink_checkParse(struct lavaInfo *lavaInfo, jsmnf_pair *field, char *fieldName);
