+++ include/plugins.h
@@ -1,3 +1,7 @@
+/** \file
+ * File containing the functions for plugin support.
+ */
+
 #ifndef PLUGIN_H
 #define PLUGIN_H
 
@@ -16,9 +20,9 @@ struct pluginEvents {
   int (*onLavalinkClose)(struct lavaInfo *lavaInfo, struct ws_info *info, enum ws_close_reason wscode, const char *reason, size_t length);
   int (*onCoglinkScheduler)(struct lavaInfo *lavaInfo, struct discord *client, const char data[], size_t size, enum discord_gateway_events event);
   int (*onDecodeTrackRequest)(struct lavaInfo *lavaInfo, char *track, struct requestInformation **res);
-  int (*onDecodeTrackParseRequest)(struct lavaInfo *lavaInfo, struct requestInformation **res, struct parsedTrack ***songStruct);
+  int (*onDecodeTrackParseRequest)(struct lavaInfo *lavaInfo, struct requestInformation **res, struct parsedTrack ***parsedTrackStruct);
   int (*onDecodeTracksRequest)(struct lavaInfo *lavaInfo, char *trackArray, long trackArrayLength, struct requestInformation **res);
-  int (*onDecodeTracksParseRequest)(struct lavaInfo *lavaInfo, struct requestInformation **res, char *songPos, struct parsedTrack ***songStruct);
+  int (*onDecodeTracksParseRequest)(struct lavaInfo *lavaInfo, struct requestInformation **res, char *songPos, struct parsedTrack ***parsedTrackStruct);
 };
 
 struct pluginEventsInternal {
@@ -28,9 +32,9 @@ struct pluginEventsInternal {
   int (*onLavalinkClose[8])(struct lavaInfo *lavaInfo, struct ws_info *info, enum ws_close_reason wscode, const char *reason, size_t length);
   int (*onCoglinkScheduler[8])(struct lavaInfo *lavaInfo, struct discord *client, const char data[], size_t size, enum discord_gateway_events event);
   int (*onDecodeTrackRequest[8])(struct lavaInfo *lavaInfo, char *track, struct requestInformation **res);
-  int (*onDecodeTrackParseRequest[8])(struct lavaInfo *lavaInfo, struct requestInformation **res, struct parsedTrack ***songStruct);
+  int (*onDecodeTrackParseRequest[8])(struct lavaInfo *lavaInfo, struct requestInformation **res, struct parsedTrack ***parsedTrackStruct);
   int (*onDecodeTracksRequest[8])(struct lavaInfo *lavaInfo, char *trackArray, long trackArrayLength, struct requestInformation **res);
-  int (*onDecodeTracksParseRequest[8])(struct lavaInfo *lavaInfo, struct requestInformation **res, char *songPos, struct parsedTrack ***songStruct);
+  int (*onDecodeTracksParseRequest[8])(struct lavaInfo *lavaInfo, struct requestInformation **res, char *songPos, struct parsedTrack ***parsedTrackStruct);
 };
 
 struct pluginSecurity {
@@ -45,6 +49,11 @@ struct coglinkPlugins {
   int amount;
 };
 
+/**
+ * Set the events that will be executed when performing a certain function.
+ * @param lavaInfo Structure with important informations of the Lavalink.
+ * @param pluginEvents Structure with the pointer to the functions.
+ */
 void coglink_setPluginEvents(struct lavaInfo *lavaInfo, struct pluginEvents *pluginEvents);
 
 #endif
